+++ simpleml/datasets/pandas_mixin.py
@@ -10,7 +10,7 @@ __author__ = 'Elisha Yadgaran'
 
 import pandas as pd
 
-from typing import Any, List, Union
+from typing import List, Union
 
 from simpleml.datasets.abstract_mixin import AbstractDatasetMixin
 from simpleml.utils.errors import DatasetError
@@ -21,22 +21,17 @@ DATAFRAME_SPLIT_COLUMN: str = 'DATASET_SPLIT'
 
 class PandasDatasetMixin(AbstractDatasetMixin):
     '''
-    "Pandas"esque mixin class with control mechanism for `self.dataframe` of
-    type `dataframe`. Only assumes pandas syntax, not types, so should be compatible
-    with pandas drop-in replacements.
+    Pandas mixin class with control mechanism for `self.dataframe` of
+    type `dataframe`. Mostly assumes pandas syntax, not types, so may be compatible
+    with pandas drop-in replacements. Recommended to implement a parallel mixin
+    for other frameworks though
 
     In particular:
         A - type of pd.DataFrame:
             - query()
             - columns
             - drop()
-            - __getitem__()
             - squeeze()
-
-        B - any other type:
-            - get()
-            - __getitem__()
-            - squeeze(
     '''
     @property
     def X(self) -> pd.DataFrame:
@@ -46,7 +41,7 @@ class PandasDatasetMixin(AbstractDatasetMixin):
         return self.get(column='X', split=None)
 
     @property
-    def y(self) -> Union[pd.Series, pd.DataFrame]:
+    def y(self) -> pd.DataFrame:
         '''
         Return the target label columns
         '''
@@ -84,39 +79,46 @@ class PandasDatasetMixin(AbstractDatasetMixin):
         if column not in ('X', 'y', None):
             raise ValueError('Only support columns: X, y, None')
 
-        if isinstance(self.dataframe, pd.DataFrame):
-            if split is None:  # Return the full dataset (all splits)
-                df = self.dataframe
-            else:
-                # query automatically returns a copy
-                if DATAFRAME_SPLIT_COLUMN not in self.dataframe.columns:
-                    raise DatasetError('Cannot retrieve dataset split `{split}` from dataframe without `{DATAFRAME_SPLIT_COLUMN}` column')
-                df = self.dataframe.query("{}=='{}'".format(DATAFRAME_SPLIT_COLUMN, split))
-            if DATAFRAME_SPLIT_COLUMN in df.columns:
-                df.drop(DATAFRAME_SPLIT_COLUMN, inplace=True, axis=1)
-        else:
-            # in case the dataframe is contained in a dict or other itemgetter enclosure
-            # expects type pd.DataFrame returned
-            df = self.dataframe.get(split, None)
-            if df is None:  # Make compatible with subscription syntax
-                df = pd.DataFrame()
-            else:
-                # copy for mutable downstream operations
-                # self.dataframe only returns a copy if type dataframe
-                df = df.copy()
+        dataframe = self.dataframe  # copy
 
-        if column is None:
-            return df
+        # choose the columns to slice from the dataframe
+        if column is None:  # All except internal columns
+            return_columns = [col for col in dataframe.columns if col != DATAFRAME_SPLIT_COLUMN]
 
-        elif column == 'y':  # Squeeze to reduce dimensionality of return
-            # inplace drop non label columns
-            drop_columns = [col for col in df.columns if col not in self.label_columns]
-            df.drop(drop_columns, axis=1, inplace=True)
-            return df.squeeze()
+        elif column == 'y':  # Just label columns
+            return_columns = self.label_columns
 
         else:  # X
-            df.drop(self.label_columns, axis=1, inplace=True)
-            return df
+            return_columns = [col for col in dataframe.columns if col != DATAFRAME_SPLIT_COLUMN and col not in self.label_columns]
+
+        return self._get(dataframe=dataframe, columns=return_columns, split=split)
+
+    @staticmethod
+    def _get(dataframe: pd.DataFrame, columns: List[str], split: str) -> pd.DataFrame:
+        '''
+        Internal method to extract data subsets from a dataframe
+
+        :param dataframe: the dataframe to subset from
+        :param columns: List of columns to slice from the dataframe
+        :param split: row identifiers to slice rows (in internal column mapped to `DATAFRAME_SPLIT_COLUMN`)
+        '''
+        if split is not None:  # Return the full dataset (all splits) - already a copy
+            # query automatically returns a copy wisth a weakref
+            if DATAFRAME_SPLIT_COLUMN not in dataframe.columns:
+                raise DatasetError('Cannot retrieve dataset split `{split}` from dataframe without `{DATAFRAME_SPLIT_COLUMN}` column')
+            dataframe = dataframe.query("{}=='{}'".format(DATAFRAME_SPLIT_COLUMN, split))
+
+        # inplace drop extra columns
+        drop_columns = [col for col in dataframe.columns if col not in columns]
+        if drop_columns:
+            dataframe.drop(drop_columns, axis=1, inplace=True)
+
+        # Last check in case any of the operations created a view or weakref copy
+        if (hasattr(dataframe, '_is_view') and dataframe._is_view) or \
+                (hasattr(dataframe, '_is_copy') and dataframe._is_copy is not None):
+            dataframe = dataframe.copy()
+
+        return dataframe
 
     def concatenate_dataframes(self,
                                dataframes: List[pd.DataFrame],