+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Dataset external file setter with validation hooks
 - Pandas changes to always return dataframe copies (does not extend to underlying python objects! eg lists, objects, etc)
 - Pandas Dataset Subclasses for Single and Multi label datasets
+- PersistableLoader methods do not require name as a parameter
 
 *BREAKING CHANGES*
 - `PandasDataset` is deprecated and will be dropped in a future release. Use `SingleLabelPandasDataset` or `MultiLabelPandasDataset` instead